backend IA/B3/requirements.txt
# Backend dependencies for AI Interview Playbook Avatar
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
python-multipart==0.0.6

# TTS Providers
elevenlabs==1.0.0
azure-cognitiveservices-speech==1.35.0
google-cloud-texttospeech==2.16.0
coqui-tts[codec]==0.27.4

# PyTorch for Coqui TTS (install separately - see README)

# Audio processing
pydub==0.25.1

# CORS support
starlette==0.35.1
